refactor(itam): mv inventory task to itam app
ref: #437
