+++ app/core/tests/functional/test_task_result/test_task_result_viewset.py
@@ -147,7 +147,7 @@ class TaskResultPermissionsAPI(
 
         self.item = self.model.objects.create(
             task_id = 'd15233ee-a14d-4135-afe5-e406b1b61330',
-            task_name = 'api.tasks.process_inventory',
+            task_name = 'itam.tasks.process_inventory',
             task_args = '{"random": "value"}',
             task_kwargs = 'sdas',
             status = "SUCCESS",
